requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
pymongo[srv]==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
flask-mail==0.9.1
dnspython==2.4.2
requests==2.31.0
